refactor(card-browser): move Grade Now to fragment
